backend/requirements.txt
django>=4.2,<7.0
djangorestframework>=3.14
django-cors-headers>=4.0
Pillow>=10.0
PyJWT>=2.8
cryptography>=41.0
requests>=2.31